Split up MetricsUpdateTaskIT so it's picked up by failsafe
